Create 1.A
Prints out matrix in each iteration showing position of organism and food particles, program will run much faster to get the required 10,000 iterations if the line to print the matrix is disabled. Output in the end gives list of positions in row, col format and current target in the last move.